Storybook: CartesianAxis - Split into simple API story and additional example
